Move segment_otsu into its own file
