+++ src/napari_mm3/__init__.py
@@ -5,3 +5,6 @@ __version__ = "0.0.6"
 from ._annotate import Annotate
 from ._channel_picker import ChannelPicker
 from ._compile import Compile
+from ._subtract import Subtract
+from ._track import Track
+from ._segment_otsu import SegmentOtsu
